cardlineShop accepts floats as price and so does SellRequestCardsAPI script

+++ src/services/sellRequestCardAPI.js
@@ -8,7 +8,7 @@ function update(
   var propertyUpdate;
   if (SellRequestCardPropertyName === "price") {
     propertyUpdate = {
-      price: parseInt(SellRequestCardPropertyValue),
+      price: parseFloat(SellRequestCardPropertyValue),
     };
   } else if (SellRequestCardPropertyName === "quantity") {
     propertyUpdate = {
